Code cleanup and Phonesky support

+++ gmail.sh
@@ -13,24 +13,26 @@ echo "" >> build.log
 echo "Updating "$INDIR" on $DATE for kitkat, lollipop, marshmallow, and nougat" >> build.log
 echo "Google Gmail add-on for 4.4.4+ (arm/arm64) (replaces stock email client)" >> build.log
 
-DIR=$(dirname "${FILEPATH}")
-FILE=${FILEPATH##*/}
-NOEXT=${FILE%\.*}
+if [ ! "$FILEPATH" == "" ]; then
+  DIR=$(dirname "${FILEPATH}")
+  FILE=${FILEPATH##*/}
+  NOEXT=${FILE%\.*}
 
-cd "$DIR"
+  cd "$DIR"
 
-if ! [ $FILE == "" ]; then
-  rm "$FILENAME"
-  mv "$FILE" "$FILENAME"
+  if ! [ $FILE == "" ]; then
+    rm "$FILENAME"
+    mv "$FILE" "$FILENAME"
+
+    VERSION=$(echo "$FILE" | cut -d "_" -f 2)
+    APIVER=$(echo "$FILE" | cut -d "_" -f 3)
+
+    cd "$BASEDIR"
+
+    echo "Version: $VERSION" >> build.log
+    echo "API: $APIVER" >> build.log
+    echo "" >> build.log
+  fi
 fi
-
-VERSION=$(echo "$FILE" | cut -d "_" -f 2)
-APIVER=$(echo "$FILE" | cut -d "_" -f 3)
-
-cd "$BASEDIR"
-
-echo "Version: $VERSION" >> build.log
-echo "API: $APIVER" >> build.log
-echo "" >> build.log
 
 ./makezipsign.sh "$INDIR" "$NAME" "$VER"
